+++ src/Addons/34thPRC_ArmourCustom/data/odst2/vasya/urban/config_weapons.hpp
@@ -35,3 +35,47 @@ class 34thPRC_ArmourCustom_ODST2_Vasya_CH252D_CustomVisor_Urban_dp : 34thPRC_Arm
 		};
 	};
 };
+class 34thPRC_ArmourCustom_ODST2_Vasya_M52_D_Urban : 34thPRC_ArmourStandard_ODST2_M52_D_Urban
+{
+	displayName="[34th] M52-D Combat Armour 'Vasya' Urban";
+	author="Over Yandere";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_cqb_urban.paa", //CQB Left
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_cqb_urban.paa", //CQB Right
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_marksman_urban.paa", //Marksman_Left
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_marksman_urban.paa", //Marksman_Right
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_vest_urban.paa", //ODST_Bracer_Left
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_vest_urban.paa", //ODST_Bracer_Right
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_vest_urban.paa", //ODST_Chest
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_shoulders_urban.paa", //ODST_Left
+		"34thPRC_ArmourCustom\data\odst2\vasya\urban\Vasya_odst_shoulders_urban.paa", //ODST_Right
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban";
+		player="vasya";
+		camo="urban";
+		variant="d";
+	};
+};