+++ src/Addons/34thPRC_ArmourStandard/data/odst2/config_weapons.hpp
@@ -4,6 +4,7 @@ class CH252D_Helmet_dp; //MA_Armor -> Marine_Helmets.hpp
 class M52_ODST_Vest_Standard_A; //MA_Armor -> M56D_Vests.hpp //Standard ODST Vest, no pouches and bobs
 class M52_ODST_Vest_Standard_B; //MA_Armor -> M56D_Vests.hpp
 class M52_ODST_Vest_Standard_C; //MA_Armor -> M56D_Vests.hpp
+class M52_ODST_Vest_Standard_D; //MA_Armor -> M56D_Vests.hpp
 class M52_ODST_Vest_Marksman; //MA_Armor -> M56D_Vests.hpp //Marksman version
 class M52_ODST_Vest_CQB; //MA_Armor -> M56D_Vests.hpp //CQB version
 class M52_ODST_Vest_Light; //MA_Armor -> M56D_Vests.hpp //Shoulderless version of M52_ODST_Vest_Standard_A