+++ src/Addons/34thPRC_ArmourStandard/data/odst2/urban/config_weapons.hpp
@@ -382,6 +382,177 @@ class 34thPRC_ArmourStandard_ODST2_M52_C_Medic_Urban : 34thPRC_ArmourStandard_OD
 		variant="medic_c";
 	};
 };
+class 34thPRC_ArmourStandard_ODST2_M52_D_Urban : M52_ODST_Vest_Standard_D
+{
+	displayName="[34th] M52-D Combat Armour Urban";
+	author="Over Yandere";
+	hiddenSelections[]=
+	{
+		"Camo1",
+		"Camo2",
+		"Camo3",
+		"Camo4",
+		"Camo5",
+		"Camo6",
+		"Camo7",
+		"Camo8",
+		"Camo9",
+		"Camo10",
+		"Camo11",
+		"Camo12",
+		"Camo13",
+		"Camo14",
+		"Camo15",
+		"Camo16",
+		"Camo17",
+		"Camo18",
+		"Camo19",
+		"Camo20",
+		"Camo21",
+		"Camo22",
+		"Camo23",
+		"Camo24",
+		"Camo25",
+		"Camo26",
+		"Camo27",
+		"Camo28",
+		"Camo29",
+		"CQB_Left",
+		"CQB_Right",
+		"Marksman_Left",
+		"Marksman_Right",
+		"LShoulderRadio",
+		"RShoulderRadio",
+		"WaistSLeft",
+		"WaistSRight"
+	};
+	class ItemInfo: ItemInfo
+	{
+		hiddenSelections[]=
+		{
+			"Camo1",
+			"Camo2",
+			"Camo3",
+			"Camo4",
+			"Camo5",
+			"Camo6",
+			"Camo7",
+			"Camo8",
+			"Camo9",
+			"Camo10",
+			"Camo11",
+			"Camo12",
+			"Camo13",
+			"Camo14",
+			"Camo15",
+			"Camo16",
+			"Camo17",
+			"Camo18",
+			"Camo19",
+			"Camo20",
+			"Camo21",
+			"Camo22",
+			"Camo23",
+			"Camo24",
+			"Camo25",
+			"Camo26",
+			"Camo27",
+			"Camo28",
+			"Camo29",
+			"CQB_Left",
+			"CQB_Right",
+			"Marksman_Left",
+			"Marksman_Right",
+			"LShoulderRadio",
+			"RShoulderRadio",
+			"TorsoPMLeft",
+			"TorsoPMRight",
+			"WaistSLeft",
+			"WaistSRight"
+		};
+	};
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_CQB.paa", //CQB Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_CQB.paa", //CQB Right
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Marksman.paa", //Marksman_Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Marksman.paa", //Marksman_Right
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Vest.paa", //ODST_Bracer_Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Vest.paa", //ODST_Bracer_Right
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Vest.paa", //ODST_Chest
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Shoulders.paa", //ODST_Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Shoulders.paa", //ODST_Right
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST2_M52_Urban";
+		camo="urban";
+		variant="d";
+	};
+};
+class 34thPRC_ArmourStandard_ODST2_M52_D_Medic_Urban : 34thPRC_ArmourStandard_ODST2_M52_D_Urban
+{
+	displayName="[34th] M52-D Combat Armour (Corpsman) Urban";
+	author="Over Yandere";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_CQB.paa", //CQB Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_CQB.paa", //CQB Right
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Marksman.paa", //Marksman_Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Marksman.paa", //Marksman_Right
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Vest_Medic.paa", //ODST_Bracer_Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Vest_Medic.paa", //ODST_Bracer_Right
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Vest_Medic.paa", //ODST_Chest
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Shoulders_Medic.paa", //ODST_Left
+		"34thPRC_ArmourStandard\data\odst2\urban\34th_Urban_ODST_Shoulders_Medic.paa", //ODST_Right
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+		"MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+		"MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST2_M52_Urban";
+		camo="urban";
+		variant="medic_d";
+	};
+};
 class 34thPRC_ArmourStandard_ODST2_M52_Marksman_Urban : M52_ODST_Vest_Marksman
 {
 	displayName="[34th] M52 Combat Armour (Marksman) Urban";